CLI/Python: Run prompts in a blocking thread
Dialoguer's interface is entirely blocking. The prompts themselves must be single-threaded because stdout can only handle one prompt at a time. By pushing them into a background thread, we allow the rest of the build to make progress while waiting on the prompt. For example, if there's a triggered request or command subprocess needed for the build as well, those can run and be handled while waiting on the prompt response.

I manually tested both the CLI and Python with multiple simultaneous prompts. Dialoguer properly queues them so the user only gets one prompt at a time.

This involved adding a new crate, because I didn't want to duplicate the prompter stuff across 2 crates. Especially since the fs frontend will need this as well, so soon it'll be 3 crates. This can't go into slumber_core because I don't want to add dialoguer and anyhow as dependencies there.
