+++ crates/console/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+authors = {workspace = true}
+description = "Utitilies for console-based Slumber frontends. Not intended for external use."
+edition = {workspace = true}
+homepage = {workspace = true}
+keywords = {workspace = true}
+license = {workspace = true}
+name = "slumber_console"
+repository = {workspace = true}
+rust-version = {workspace = true}
+version = {workspace = true}
+
+[package.metadata.release]
+tag = false
+
+[dependencies]
+anyhow = {workspace = true}
+async-trait = {workspace = true}
+dialoguer = {workspace = true, features = ["password"]}
+slumber_core = {workspace = true}
+slumber_template = {workspace = true}
+slumber_util = {workspace = true}
+tokio = {workspace = true, features = ["sync"]}
+tracing = {workspace = true}
