+++ crates/python/Cargo.toml
@@ -22,10 +22,10 @@ crate-type = ["cdylib"]
 
 [dependencies]
 async-trait = {workspace = true}
-dialoguer = {workspace = true, features = ["password"]}
 indexmap = {workspace = true}
 pyo3 = {version = "0.27.2", features = ["experimental-async", "indexmap"]}
 slumber_config = {workspace = true}
+slumber_console = {workspace = true}
 slumber_core = {workspace = true}
 slumber_template = {workspace = true}
 tokio = {workspace = true, features = ["rt-multi-thread"]}
